requirements.txt
html2docx
beautifulsoup4
pandas
python-docx
openpyxl